+++ InterlocutorComercial/InterlocutorComercialRN.h
@@ -8,11 +8,11 @@ private:
 public:
 	InterlocutorComercialRN(string nombreArchivo = "");
 	~InterlocutorComercialRN();
-	bool ExisteInterlocutorComercial(long dni);
-	InterlocutorComercial BuscarInterlocutorComercial(long dni);
+	bool ExisteInterlocutorComercial(int dni);
+	InterlocutorComercial BuscarInterlocutorComercial(int dni);
 	bool AltaInterlocutorComercial(InterlocutorComercial& interlocutorComercial);
 	bool ModificaInterlocutorComercial(InterlocutorComercial& interlocutorComercial);
-	long CantidadInterlocutoresEnSistema();
+	int CantidadInterlocutoresEnSistema();
 	string getNombreArchivo();
 	vector<InterlocutorComercial> VectorInterlocutores();
 	void setNombreArchivo(string nombreArchivo);
